Enable serial port debug print inside the kernel
In preparation for removing video debug messages from the kernel,
an alternate debug print method was needed. The serial port was
chosen.  The driver comes from SYSLINUX, nearly verbatim.  Only
the ordering of parameters in the outb() function had to be changed
and their macro made that really easy to do.
Good work Mr. Anvin, that is is a really portable driver.