Fixed problems with __PACKED and outportb
The function outportb() was renamed to outb() but because I don't
have dependency files set up yet, the compilation passed even with
the old definition until I did a full rebuild. There was a similar
problem with the __PACKED macro being moved from the end of the
struct to the beginning, which also didn't compile properly.