New script for creating ISO images
The previous script, makeimage.sh has been superceded by makeiso.sh.
The old script made a disk image, and then generated an ISO which booted
from the image.  This approach was problematic (it could never find the
kernel).  The new method is to create an ISO image straight off, with
all of the files in the ISO rather than in an emulated hard drive image.
This seems to work much better under SYSLINUX, and it can now find the
kernel file.