Added nostdinc to CFLAGS on JOEL4 kernel
The nostdinc flag tells GCC to not look in any of the normal include paths for
header files.  When building a kernel, sometimes the local definitions are a
bit different than the standard definitions for a function, and including the
standard header files will give warnings and sometimes even conflicts.  This
option allows me to reuse function names and even header file names without
any conflicts.