kernel: force \r to be printed with \n on serial
So the line endings show up correctly in Windows when viewing the output from
the serial port from Virtualbox, force DOS line endings.  I know this doesn't
work right if the k_printf already contains \r\n or \n\r, but since I'm the
only one writing kernel print messages, it works for me.  Of course this can
be changed in the future if needed.