joel4: adding in-kernel dynamic memory allocation
Used for linked lists and other situations where dynamically allocated
memory is the way to go.  The kernel has a small local heap from which this
memory is accessed.  The heap is statically allocated, meaning the physical
memory for the entire kernel heap is always available.  For that reason, we
try to keep it as small as possible without starving the kernel.