joel4: allocator for aligned pages in the kernel
This serves up 4KB pages, aligned to 4KB.  There are a very limited number
of these pages, and are meant for a very specific purpose (actually 2).
First, for creating the page directory for a new task, and second for
temporarily mapping page tables for new tasks.  The page directory
remains allocated for the life of the task, whereas the page tables should
be freed after the kernel has done its mapping.