joel4: improvement in page mapping for root task
Not 100% yet, but kernel no longer triple faults.  We can now load
pages, and start execution.  There is a problem still with logical
mappings of physical pages, in part because this is for the root task
which is a special case.  We need to pick a logical execution address
and stick to it.