joel4: can make syscall
The root task pages have been mapped, and we can now access code in
the first userspace task.  To start, we must test the syscall
mechanism, which allows userspace code to call a kernel function.
The call works (we can debug print a character) but the return path
(sysexit) doesn't work yet.