joel4: automatic build versioning
Added a shell script and a separate header file to keep track of the kernel
version.  The build number is automatically incremented before each build,
and the header is updated with the new number.  When the file is compiled,
the define is made available to the kernel, and is printed on boot.

Signed-off-by: Joel Nider <joel@definiware.com>