GRUB2 bootloader support
Added GRUB2 as an option for a bootloader (alongside SYSLINUX/ISOLINUX).  This
was done primarily to gain the option of adding a kernel module (init.cpio).
I couldn't figure out how to do this with ISOLINUX, even with their extensive
5-line wiki.  GRUB2 is also a major pain, but at least I already know it.
Seriously - who thought 'grub-mkrescue' was a good name for burning an ISO?
I could understand if this -also- made an ISO, but it seems that it is the
-only- way to make an ISO.

Signed-off-by: Joel Nider <joel@nider.org>