joel4: added boot info and init.cpio support
Information needed for system initialization by the root task is now passed
by the kernel as a parameter to the root task.  This is a special case, and
the format should probably be changed again in the future once there is a
more standard main for a new task (like argc/argv).

Also, the init.cpio file is loaded by the boot loader into physical memory.
Now the kernel copies it to a new location mapped into the logical space of
the root task, so the root task can use the file.

Signed-off-by: jnider <joel@nider.org>