joel4: fixed static variables problem
The loader script was missing the static variables from the root task
and putting them at the wrong address. I can only assume that the section
name is different, but I wasn't able to determine the correct section name.
This is a work-around by selecting everything that isn't .text (code) and
putting it in the data section.  This is exactly what I want anyway, but it
would be nice to know what the correct section name is, and why my previous
solution didn't work.

Signed-off-by: jnider <joel@nider.org>