joel4: preparation for release mode builds
Modified the makefiles and tools.mk to allow for separate debug/release
build configurations.  The debug mode adds the -g flag to gcc compilations
which includes DWARF2 debugging information in the object. Also, the -O0
flag is specified, even though its the default.  In the release configuration,
-O2 should be specified, but currently cannot be set due to the 'ldr' bug
(possibly in gcc, possibly in my code).

Signed-off-by: jnider <joel@nider.org>