boot: fixed the paths to the init modules
boot.txt contains a list of files to be loaded by the boot task. These
files are contained in a cpio archive which is sensitive to path names.
Currently the cpio archive is built from the source tree, which contains
certain paths, and these paths are then carried over to the cpio file.
It just means that the paths must be specified in order to find the files
correctly when parsing the cpio archive.

Signed-off-by: jnider <joel@nider.org>