JOEL4: added cloc target
Can now use cloc to count the number of lines of code in the kernel.
There is little chance of me remembering all of the correct options
to all of the command line tools required in this (and my other)
project(s), therefore I made it a 'make' target.
This of course requires the cloc tool to be installed.

Signed-off-by: Joel Nider <joel@nider.org>