JOEL4: x86 segments should be 16-bit values
When loading the task register with LTR, a 16-bit segment must be used.
A bug in the macro led the LTR instruction to try to load from a 32-bit
register, which failed during assembly.  This bug was only seen when
compiling with O1 or greater, since a memory location was used with O0
which does not exhibit the same behavior.

Signed-off-by: Joel Nider <joel@nider.org>