JOEL4: fix for LTR assembler syntax bug
When compiling with O1 and higher, the compiler emits an assembly sequence
that loads the task register (LTR) from a register rather than a memory location.
Since I was using a 32-bit value, the compiler naturally tried to load the LTR
from a 32-bit register.  However, the instruction can only load from 16-bit
register. Under O0 this was not seen because the memory syntax seems to be more
flexible, and the assembler was able to reference the memory as a 16-bit
location.

Signed-off-by: Joel Nider <joel@nider.org>