JOEL4: not returning a value in taskListFindByID
The compiler should have caught this one. I was trying to return from a function
without returning a value. This caused an infinite loop while creating the root
task. I should probably add a check for this case in the createTask function
since this still may happen if all task ID's are really in use (not likely, but
the kernel should be able to detect this case.

Signed-off-by: Joel Nider <joel@nider.org>