JOEL4: Added KIP mapping for root task
The kernel information page can now be retrieved in the root task by calling
the L4_KernelInterfacePage system call. The information is not yet complete,
but the mechanism is in place.

Removed the boot info hack which was causing problems when compiled with O1.
Apparently the stack looks different, and this hack was only correct for the
O0 stack frame.

Signed-off-by: Joel Nider <joel@nider.org>