JOEL4: added invalid sysenter handler
It's called syscall_handler but its actually for the SYSENTER mechanism. I added
rudimentary bounds checking on the array to make sure we are not trying to jump
to a function that doesn't exist in the table. I call this rudimentary because
the index is hard-coded, and will not update automatically when more handlers are
added to the table. I don't expect many handlers to be added to the table over
time (should be fairly static based on the L4 API and possible additions) however
it's always nicer when numbers are defined in a single place wherever possible
and not spread out across several files, including an assembly file in this case.

I also found the bug that was causing the jump to an invalid location (which is
why I added the bounds checking in the first place). The bug was caused by my
trying to update the eax register with an input constraint that should read from
a constant. I tried setting the constraint to "aI" but apparently that is not
correct. I still don't know if this is possible, but I ended up doing it the
old fashioned way (mov $constant, %eax) which seems to work fine. I will have to
learn more about these inline assembler constraints, and possibly update the
code if there is a better way to do this.

Signed-off-by: Joel Nider <joel@nider.org>