JOEL4: removed BootInfo as parameter to root task
Originally I tried to stuff the BootInfo into the stack of the root task
forcefully from the kernel. It was a horrible hack that finally fell apart
once I built with -O2, since the stack frame looks different. This caused
me to look for an alternative, which led me to discover that there is room
in the KIP for exactly this purpose.  How nice - I now use the KIP the way
it was designed, and saved myself a lot of trouble.

Signed-off-by: Joel Nider <joel@nider.org>