KOS: missing $ in gdt constant (asm)
Even though the KOS kernel is no longer being developed, it is nonetheless
being used as a reference since it is posted on github and shows up in
searches. In this light I feel it is important to fix any bugs as to not
mislead people accidentally who may rely on this code without realizing its
status.

Constant values in AT&T style assembly should start with a $.
Thanks to Lars Gehlert for pointing out this bug.

Signed-off-by: Joel Nider <joel@nider.org>