JOEL4: split up user and kernel space L4 interface
Userspace drivers must be able to access the L4 kernel functions for IPC
 etc. To do so, they must be able to link with a library containing the
stub functions. Currently these stub functions are contained in the same
file as their kernelspace counterparts. That doesn't work well when trying
to isolate kernel functionality from userspace functionality since it is
not possible to compile the file with only userspace definitions. Thus,
all of the of the userspace functions have been moved to a separate file
which can be used as a library for linking against userspace drivers and
other applications wishing to access the kernel directly.

Signed-off-by: Joel Nider <joel@nider.org>