map boot modules into root task address space
The bootloader loads modules from the boot device into memory and
leaves them there for the kernel. The kernel uses one of the
modules as an initial ramdisk in order to find the root task. Once
the root task is loaded into memory, it then takes over responsibility
to load the rest of the system. It thus needs access to the same
archive which contains itself along with other drivers. The kernel
must map the memory pages containing the module into the app's
address space, and point to it through a special header which is
mapped at a known location.

Signed-off-by: Joel Nider <joel@nider.org>