Move L4 interface header to user include directory
This file is shared by both kernel and user space programs. However,
it is not a good idea for user space programs to include the general
directory containing all of the internal kernel headers, since there
may be conflicts with filenames or function names. In any case, the
user space programs have no need to know anything about the kernel
internals, and only need this one interface file. Therefore, I am
moving it to a neutral, shared location that is isolated, yet
accessible by all.

Signed-off-by: Joel Nider <joel@nider.org>