Move CPIO header into user space
The user space code was borrowing the kernel header, which is not a good idea
anymore since the implementations have diverged. So to avoid conflicts, I made
a copy of the header and put it in a user space only directory. Eventually,
this should be moved out into a library that can be linked from any user space
application. Eventually.

Signed-off-by: Joel Nider <joel@nider.org>